orbit_consensus_propagation/requirements.txt
matplotlib>=3.5.3
numpy>=1.21.6
pandas>=1.3.5
pyephem>=9.99
pymoo>=0.6.1.1
pytictoc>=1.5.3
PyYAML>=6.0.1
Requests>=2.31.0
scipy>=1.7.3
seaborn>=0.12.2
